Add Python compatible Lambda Runner
Python Lambda Runner uses the child_process module to spawn the sls
process. This code is based on serverless-offline-python. It receives
the serverless directory, the handler path, the handler method and the
payload. The serverless directory will indicate where the sls command
should be invoked. In this case the handler function is configured like
the following example:

```
functions:
  example:
    handler: example/handler.example
    timeout: 10
    runtime: python3.6
```

The environment is inherit from the main process.