Report Exception in server response (#100)
Lots of code diff here but all I really did is wrap this piece of code into a try catch.
When an exception was thrown, the error was not clear at all. The server returns 500 and the console would show nothing at all.
For example, on line 264, when no auth is passed
On line 271, `validate` throws an exception when the schema itself is in invalid.
Or on line 303, when invalid query type. 

This change makes the error clear, and directly into the GQL response.